crash: KernelCollector: Improve hung task sigs by ignoring boring functions
The top function in a hung task is often a "boring" function. That
means that with hung tasks we don't get a very good human readable
name. Let's add a list of boring functions to ignore so we can pick
something a little better.

NOTE: this list of boring functions is _different_ than the list used
for magic signatures. Notably it is a smaller list and also doesn't
match whole classes of functions based on prefix or suffix. This is
probably OK. In general it's should be OK to have signatures based on
different logic. There may be cases where one signature will give a
great human readable string and the other won't. Specifically, having
one that is conservative about boring function (only removes ones that
are _truly_ boring) and one that is more liberal (removes things that
are _probably_ boring) gives us two options.

BUG=b:309669058
TEST=Unit tests

Change-Id: If7f8b6b3abac8b3e4498cbcdbff9c67e276eca17
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5055610
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>