power: Add kernel_default sleep mode preference
Add the `kernel_default` sleep mode preference.  With this preference
set the system's initial mem_sleep mode at runtime will be used as the
method instead of the config decision.  With this configuration the
kernel's default value can be used for the sleep mode.

This will allow Flex to defer the selection of the mode dynamically to
what the kernel supports for the current device.

BUG=b:275614676
TEST=`cros_workon_make --board=reven power_manager --test`
TEST=Test on a reven machine with s2idle as the kernel default.\
  echo kernel_default > /var/lib/power_manager/suspend_mode \
  restart power_manager \
  powerd_dbus_suspend \
  Check powerd logs and confirm confirm `/sys/power/mem_sleep` is \
  `s2idle` after resume.

Change-Id: I20195ee7877a61a4b27c2c889f04fa6720b53d2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5056246
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Jeffery Miller <jefferymiller@google.com>