shill: network: avoid modifying event handlers vector during iteration
The tethering attempt is completed in place when the Network
`OnConnectionUpdated()` event is called, and this leads to the
TetheringManager attempting to add a new event handler in the Network
while the iteration of event handlers calling `OnConnectionUpdated()`
is still ongoing.

Modifying the vector makes the existing iterator be reset, and we end
up crashing while attempting to dereference it.

Making a copy of the event handlers vector and iterating that one
instead ensures we don't touch the original one.

Fix suggested by Andrew Lassalle <andrewlassalle@chromium.org>

BUG=b:312103659
TEST=Enable hotspot in brya/fm350 with a custom DUN-only APN

Change-Id: I236e359cf22fee426166bfd23b2b1c1012735924
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5054639
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>