shill: portal: simplify HTTPS probe result tracking
This patch removes several fields of PortalDetector::Result for tracking
the result of the HTTPS probe and replace them with HttpRequest::Error.

The HTTPS probe is now considered successful if the request completed
without an error.

BUG=b:309175584
TEST=FEATURES=test emerge-$BOARD shill. Flashed hatch, connected to
captive portal, verified that sign-in notification is shown and that
portal redirection is detected in /var/log/net.log.

Change-Id: I4a00e8a03f0a897a15cb10db04f03dcb67eb43db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5056097
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>