cryptohome: Add BoundAuthSession
This attempts to fix the "in use session bound to a callback that never
gets called" by disallowing the binding of InUseAuthSession objects to
callbacks and instead making a BoundAuthSession wrapper that
incorporates a timeout that will de-use the session if the timeout
triggers. The idea is to guarantee that we'll never accidentally keep a
session in-use (blocking all sessions for a user) forever.

Unfortunately this also somewhat undoes the trick of binding an in-use
object to the on_done callback passed to AuthSession functions. This
works fine so long as those functions just do a single unit of work, but
if they schedule additional work bound to themselves (which auth session
unfortunately does) then they can get into a state where the they keep
doing new work but the session is no longer in use.

So this will likely require _another_ followup CL to really fix that, by
making a better mechanism for AuthSession functions to properly forward
their in-use state between their own internal callbacks.

BUG=b:311711733
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ie910fb357c7f4e6309b52a265412b6f242cbf342
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5064639
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Howard Yang <hcyang@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>