metrics: Add support for obtaining logs from debugd
Some sysfs and debugfs files are readable to root only. Because of that
we can't read them directly from the metrics daemon, which runs with
reduced privileges.
To address that implement debugd_reader class. It uses dbus to fetch
a log by calling the debugd "GetLog" method. The return value is a string.
The caller is responsible for parsing it.

BUG=b:309423643
TEST=P2_TEST_FILTER="DebugdReaderTest.*" cros_workon_make --test --board=dedede metrics

Change-Id: I53cf64f62992b84aa8efdc02497bd22dcf576939
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5098834
Tested-by: Kornel Dulęba <korneld@google.com>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Kornel Dulęba <korneld@google.com>