cryptohome: Allow cancellation of callbacks against AuthSession
This is an attempt at solving the "what if we time out an
InUseAuthSession but then one of its callbacks returns?" problem
introduced by adding time outs to InUse. Basically: we just cancel any
existing callbacks.

This isn't perfect, because it would probably be ideal if we could call
the actual on_done callbacks, rather than just never triggering them.
But unfortunately that can't really easily be done in a transparent way,
and it would require basically rewriting all of the async functions in
AuthSession to use a slightly different idiom with a more explicit
cancellation. This change is obviously much simpler.

BUG=b:311711733
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I5b994151cb2e775e3855d77ad428897fc1bf27b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5074512
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Howard Yang <hcyang@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>