shill: Reset network to default MTU upon Network::Stop
In the previous code MTU is only restored when a configuration is ready
for the new network. This patch proactively reset MTU upon network stop,
to prevent the possibility that a small MTU from previous network
blocking the initialization procedure on new network. (Possible for EAP,
which happens before DHCP).

BUG=none
TEST=unit

Change-Id: Ia174800ca079268056cce8520ca1d117c53c96d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5076017
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Taoyu Li <taoyl@chromium.org>