cryptohome: Fix response handling in StartMigrateToDircrypto
This was broken by crrev.com/c/5018597 which made the implementation
async but which did not properly relocate the response to the mount
thread. This change should fix this by applying the same pattern we use
on all of the other async implementations (which the original change
probably should've done to start with).

BUG=b:312760999
TEST=FEATURES=test emerge-$BOARD cryptohome
TEST=tast run cryptohome.VaultMigration.fscrypt_v2

Change-Id: I571210e6bbdae20c6737954a7939df7667bd460f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5078079
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>