shill: cellular: Skip updating custom APNs with modb data
When using the APN revamp UI, the custom APNs should not be updated with
values from the modb when the APN names match. This is because the APN
revamp gives the user full control over the APN fields, and it now
supports the fields(ip type, auth, APN type) that the old UI was
missing.

BUG=b:314192325
TEST=Add callbox APNs to the modb, and then enter custom APNs matching
those APNs. Verify that the custom APNs don't get updated with values
from the modb when fields are empty.

Change-Id: Ie26d7c418f57df431fe7a43059bb25a65c26c77f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5079590
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>