camera: Move and rename AutoFramingSM to FramingSM
Since AutoFramingStreamManipulator will be used to handle manual zoom
and super resolution photo taking, so we rename it to
FramingStreamManipulator to reflect the new use case. The stream
manipulator is also moved to common/ as it will be used on all devices
in the future.

This CL also renames some auto-framing specific constants and functions
to be different from common framing functions.

BUG=b:314267048
TEST=auto_framing_tests

Change-Id: I89f0986f0507983905b1919205399d6d6a07bc44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5084076
Commit-Queue: Kam Kwankajornkiet <kamchonlathorn@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Tested-by: Kam Kwankajornkiet <kamchonlathorn@chromium.org>