login: Avoid starting khugepaged on <8GB devices
We don't use huge pages on <8GB devices (see CL:3886363, CL:3117425,
CL:3138171), but we still enable THP support in kernel by default. As a
result, khugepaged is running even if it is never used. As a side
effect of starting khugepaged, vm.min_free_kbytes is forced to be
not less than 132MB. So e.g. on Octopus with 4GB RAM, vm.min_free_kbytes
is 132MB instead of the default ~7.8MB, i.e. there are >120MB of memory
neither used by khugepaged for its needs, nor available for normal
usage.

To avoid this waste and regain those 120+ MB of available memory,
prevent starting khugepaged on <8GB devices where it is not really used.

CL:5084648 prevents starting khugepaged by default, while this CL ensures
starting khugepaged but only if we have >=8GB memory.

These changes decrease vm.min_free_kbytes from 132MB to its default value
~7.8MB on 4GB devices. Low and high watermarks are also correspondingly
decreased from 165MB and 198MB to ~11.5MB and ~15MB respectively.

In fact this increases available memory even more than by ~120MB. Since
the max number of per-CPU free pages is set to the low watermark (165MB)
by default, so e.g. on Octopus with a dual-core CPU, there are up to
82.5MB of per-CPU pages not available for allocations on another CPU.
This change reduces that to ~11.5MB (~5.7MB per CPU).

On the other hand, the reduced watermarks may, in theory, make kswapd
less responsive to memory pressure. If we see an increase of memory
stalls because of that, we may want to increase the watermarks via
increasing vm.watermark_scale_factor.

For now verified on Octopus with 5.15 kernel with the following tests:

- ui.TabSwitchCUJ: under mild memory pressure conditions, disabling
  khugepaged reduces memory stall rate 2x or more (thanks to increased
  available memory and thus reduced swap usage).
  See b/306228767#comment20

- multivm.Lifecycle.host_with_bg_arc: under severe memory pressure
  conditions, disabling khugepaged neither increases nor decreases
  memory stall rate. See b/306228767#comment21

BUG=b:309475612, b:306228767
TEST=khugepaged is not running on a 4GB device (e.g. octopus)
but is running on a >=8GB device (e.g. brya)

Change-Id: I7da0ff2e09f00e70fcc3ea43bb9feabdacda3bcb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5079480
Tested-by: Dmytro Maluka <dmaluka@chromium.org>
Commit-Queue: Dmytro Maluka <dmaluka@chromium.org>
Reviewed-by: Lepton Wu <lepton@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>