cryptohome: Drop a lot of auth_factor_ filename prefixes
We started adding files to cryptohome/auth_factor/* without prefixing
auth_factor_ to every filename. This goes back and applies that to a
bunch of the existing filenames.

Unfortunately this is a bit of a conflict magnet and git blame ruiner
but that's just the way file shuffles are.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: If4f35fe17648b2b992be67afbdd87bec09b8986e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5082270
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>