metrics: Increase lock grab attempts to five
For the non-blocking lock case, we only try a maximum of two attempts to
grab the lock before giving up. The default wait in between each
attempts is 20 milliseconds, so we give up after waiting on 20ms.

The current behavior of two attempts is too aggressive and causing some
tast tests to fail since the lock could not be grabbed.

Increase the number of attempts to five before giving up, so we wait a
maximum of 100ms. Relaxing the behavior should allow the lock to be
acquired more often.

BUG=b:312758730
TEST=unit tests

Change-Id: I04a13bcb7a3edb873dbcf27b82335206d3779662
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5085676
Tested-by: Kendrake Tsui <kendraketsui@google.com>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Kendrake Tsui <kendraketsui@google.com>