shill: wifi: remove the country member in WiFiProvider
The atomicity of WiFiPhy info and country code cannot be guaranteed as
these two information are conveyed from Kernel to user space in two
separate nl80211 messages.

Right now, the country is only used in hotspot when the WiFi is
non-self-managed solution. It skip setting the region code to WiFi if
the country cached is same as the one indicated in cellular MCC. To
ensure the atomicity of these two information requires overkilled
for little benefits.

This CL remove the country cache in Shill and request region update
whenever starting hotspot in non-self-managed solution.

BUG=b:314204413, b:309964789
TEST=FEATURES=test emerge shill
TEST=on nissa/gothrax tast run ${DUT} network.HotspotUpdateConfigurationWhenHotspotOn.tmobile

Change-Id: I1aa6d8677c334a058dd7054b82f63d92948dabdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5085679
Tested-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>