resourced: Add callback for flag change
Add the ability to provide a callback that is invoked when a given
feature flag is modified. This callback can be used to keep state in
sync with feature flags, without the need to poll the flag.

To avoid the potential for deadlocks due to invoking the callback under
the global feature mutex, this change reworks synchronization. By
requiring that all features be registered before initializing the
feature manager, the only synchronization needed is atomic booleans for
the feature enable flag.

BUG=None
TEST=manual verification

Disallow-Looks-For-Green: True
Change-Id: I329712256d705a7d5373b0e31f80199549b74770
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5087053
Reviewed-by: Stanisław Kardach <skardach@google.com>
Commit-Queue: David Stevens <stevensd@chromium.org>
Reviewed-by: Shin Kawamura <kawasin@google.com>
Tested-by: David Stevens <stevensd@chromium.org>