metrics: introduce a new structured metric to collect USB device info for Crostini
Define a new structured metric to collect basic USB device info on
vendor, product, and class, as well as a boolean for success of
connecting the device to Crostini.

Design Doc: go/crostini-usb-peripherals-pop-list

The metrics will be recorded in platform2/concierge during the request
to attach a device to Crostini.

BUG=b:315149244
TEST=emerge-volteer metrics

Change-Id: I307342a2694036f98ba1d384057354bfa8966690
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5096374
Reviewed-by: Jong Ahn <jongahn@chromium.org>
Tested-by: maciek swiech <drmasquatch@google.com>
Commit-Queue: maciek swiech <drmasquatch@google.com>