cryptohome: Add UssManager for managing shared Uss objects
This change does not use the manager for anything yet. What exactly the
class interface should look like is something of an interesting question
by itself so this just implements the manager class to start with.

One particularly important aspect of this is that we don't want any
session to just be able to look up the DecryptedUss for a user from a
map: we want them to have the keys they would need to actually load it
in order to access it. This makes the interface and state more complex,
it can't just be a simple "give username, get DecryptedUss".

BUG=b:306423754
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I4b43be09304bc1a3ca2b8ef123a4888ceb3d6f60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5101928
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>