shill: add the NetworkMonitor class
This CL adds the NetworkMonitor class that we plan to integrate the
PortalDetector and the CapportClient in the future. In this CL, the
NetworkMonitor only contains the logic that Network interacts with
PortalDetector now. We will add CapportClient into it in the following
CL.

To avoid the temporary duplication, this CL will be merged with the
hook-up CL (crrev.com/c/5098269) together.

BUG=b:305129516
TEST=FEATURES=test emerge-brya shill

Cq-Depend: chromium:5098269
Change-Id: I9f13b03a582686daa2038cabaadc8b41ad224d8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5070073
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>