vm_tools: concierge: emit structured metric when USB device is connected to Crostini
As we make USB passthrough to Crostini more permissive, we would like to
understand what sorts of devices are being passed through, and
potentially be able to identify regressions that cause certain devices
to no longer work properly.

To this end, we emit a structured metrics as part of concierge's
AttachUsbDevice flow - since this is a point where we have access to the
properties of the USB device as well as know which guest a device is
being passed to.

This change enables logging for only Crostini/termina guests, but paves
the path for other guests to enable logging as well.

BUG=b:315149244
TEST=emerge-volteer vm_host_tools
TEST=cros deploy dut vm_host_tools
TEST=connect USB peripheral to Crostini, check /var/lib/metrics/structured/events

Cq-Depend: 5118143
Change-Id: Ic787f37c55773c7930c70348be5a0fe6752875d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5101348
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Commit-Queue: maciek swiech <drmasquatch@google.com>
Reviewed-by: Jong Ahn <jongahn@chromium.org>
Tested-by: maciek swiech <drmasquatch@google.com>