camera: run camera_device_status_change inline when possible
camera_device_status_change currently always post a task to the camera
module thread for processing the status change. This makes it difficult
in fake HAL which runs the configuration change processing on the same
thread, but want to make sure the device status change is processed
synchronously while the fake HAL configuration is changed.

Change the callback so if the caller is already running on the right
thread, just call the CameraDeviceStatusChange immediately instead of
posting a task.

BUG=b:296013012
TEST=camera.CCAUIStress* camera.CCAUISmoke*
TEST=camera.CCAUIQRCode and check the 300ms delay is there in log

Change-Id: I7e22f20e8d673bc6a7b98b3b8a9bd6c2ac6c1185
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5102374
Reviewed-by: Shik Chen <shik@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>