shill: report suspend readiness asynchronously
See CL:5100673 for more details. There's a slight behavioral
change here in that suspend actions completion won't wait for
the return value from powerd now, but we dropped it anyway, so
it's not clear there was much point in waiting to begin with.

Additionally, since this was already run in a callback context
and that callback didn't take its own continuation, it was already
signalling that callers have no way to be called back when the
operation is completely finished. If we want to be able to do this
we should thread another callback through that pathway in Manager.

BUG=b:235259071
TEST=deploy to device, ensure that suspend readiness is reported
  and that report is logged

Change-Id: I31a089876386525c2817105e58229ca0f2f89eb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5106749
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Douglas Anderson <dianders@chromium.org>