shill: Ignore GroupFinished with unknown path
This CL changes GroupFinished logic in P2PManager to ignore signals
with unknown interface_path. We should not assume it belongs to
pending transaction if we are not able to verify it, since it may
be a fake signal (eg. lost in the past). It is better to wait for
GroupStart timeout than misinterpret received GroupFinished event.

BUG=b:301049348,b:308081318
TEST=FEATURES=test emerge-guybrush shill

Change-Id: I807e3998d2414d6751721fb8243654147ddfd5e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5111183
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Sławomir Rosek <srosek@google.com>
Reviewed-by: Kevin Lund <kglund@google.com>
Commit-Queue: Sławomir Rosek <srosek@google.com>