shill: wifi: Trigger RestartFastScanAttempts
When the device is connected to an AP and that AP disappears, the driver
will sometimes disconnect with reason code 4. In this flow,
wpa_supplicant first reports an endpoint removal update before reporting
a CurrentBSS change event. When shill receives the endpoint removal
update, it triggers the disconnection flow and resets current_service_.
When shill receives the CurrentBSS change event afterward, it'll enter
HandleDisconnect and exit early without hitting RestartFastScanAttempts
at the bottom of the function. If, for whatever reason, the single scan
triggered after HandleDisconnect fails, the device will remain idle and
not continue attempting to discover/connect.

This CL replaces Scan with InitiateScan to trigger
RestartFastScanAttempts to cover this edgecase, so the device does not
have to wait for 60 seconds to trigger a new scan and find the available
APs.

BUG=b:306732236
TEST=Steps:
Note: Used Brya/Marasov with MTK WiFi and OpenWRT AP for
testing since this disconnection secnario is more reproducible on
devices with MTK WiFi when connected to our
OPENWRT[Ubiquiti_UniFi_6_Lite]
1- cros_workon --board=brya start shill
2- cros_workon_make --board=brya shill --test [[ ALL PASS ]]
3- emerge-brya shill
4- cros deploy localhost:2200 shill
5- Ran the test wifi.SSIDSwitchBack for 100 iterations [[ ALL PASS ]]
6- Checked net.log to make sure the RestartFastScanAttempts is triggered
and the AP is found as a result of that. The devices reconnects to the
AP in ~12 seconds.
7- Ran "FEATURES=test emerge-brya shill" with this change and the test
FastScanBSSLost [[ PASS ]]
8- Ran "FEATURES=test emerge-brya shill" without this change and the test
FastScanBSSLost [[ FAILED ]]

Change-Id: If92fd6274c5ae6842e93850465ca5e496190be81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5118419
Commit-Queue: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Arowa Suliman <arowa@chromium.org>