userfeedback: snapshot sensor raw data
To better understand tablet mode issue, collect the status of the iio
sensors, including the accelerometers and lid angle.

We can use that information to triage issues related to sensors
themselves or with message passing of the tablet status.

The collection does not change the privacy envelop, as the raw sensor
data is volatile and change frequently.

BUG=b:290272841
TEST=Check all raw sensor attributes are collected.

Change-Id: I8a47187e2fa9dff7b74ed5e33bc30045aa5630c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5113107
Auto-Submit: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Xiangdong Kong <xiangdongkong@google.com>
Tested-by: Gwendal Grignou <gwendal@chromium.org>