lorgnette: Set device_uuid for IPP-USB devices
Returned ScannerInfo objects are supposed to have a device_uuid field
that lets the caller determine when two connection strings refer to the
same physical scanner.  IPP-USB scanners using eSCL have an internal
UUID that can be used for this purpose, but to start with, just return a
random UUID.  This will allow other USB entries to be de-duplicated
against the IPP-USB device within the same session.

BUG=b:311196232
TEST=lorgnette_cli discover --show_details

Change-Id: I8daeaf06881f01b94f68864816b7c985d3e0b7c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5136115
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>