lorgnette: Disable hanging pixma options
The pixma backend hangs when retrieving the value of various advanced
read-only options.  This causes lorgnette to hang during scanner
discovery when it attempts to read the current set of option values.
This is reproducible in scanimage, so it does not appear to be a problem
with lorgnette.  Until we can resolve this problem upstream, disable all
of the problem options by forcing them to be inactive.

BUG=b:274860636
TEST=lorgnette_cli discover

Change-Id: I21ee3ce6353523046bb38cd7b6a681ea82ef31dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5132588
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>