cryptohome: Pass metadata to CreateCredentialVerifier
CredentialVerifiers need to be able to reconstruct auth factors. This
means that the auth factor metadata will need to be provided to the
credential verifier. Currently smartcard factor is the only factor that
have both metadata fields and a credential verifier, and it passes
metadata by reconstructing from the AuthInput. Taking auth factor
metadata as a CreateCredentialVerifier will be simpler to support these
kinds of factors.

A bit of refactor to pass around the metadata is needed to make this
change.

BUG=b:312628857
TEST=cryptohome unittests

Change-Id: Ie6ddf7d84638ba7927b5dbcc82a04d130c04fe96
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5117078
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>