shill: Migrate to NetworkConfig for ArcVpnDriver
Migrate GetIPv4Properties and GetIPv6Properties to GetNetworkConfig for
ArcVpnDriver. In the base class VPNDriver, GetIPv4Properties and
GetIPv6Properties are made not pure virtual because they are removed
from the sub-class ArcVpnDriver, but they should still not be reached
from the base class.

BUG=b:307855773
TEST=cros_workon_make --board=${BOARD} shill --test
TEST=tast run ${DUT_IP_ADDRESS} network.ARCVPNConnect

Change-Id: I464262f21747f7c63ad0b22316e2c4dbee7e6bb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5116921
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Zikai Chen <chenzikai@google.com>
Commit-Queue: Tim Bain <tbain@google.com>