hiberman: Witness the ActiveMount type with the Drop trait
The code had been written in a very OOP frame of reference where
the Drop trait was viewed as a virtual class however this is a
incorrect interpretation of the function of a Rust trait. Rather,
the Drop trait is a contract upon the Type 'ActiveMount' where its
life-time is witnessed by lexical scope and a `drop()` call shall
be automatically inserted at compile-time once it is out of scope.

Therefore inline the umount of the volume into the Drop contract
upon the type of ActiveMount. This ensures a umount will always
be dispatched when ActiveMount's life-time has ended. In by doing
so we reveal deeper issues around the lack of type coupling between
a ActiveMount's value level side-effects and its users. This is next
to be addressed towards the ultimate goal of having the type system
ensure the mount life-time is coupled to that of the value instantiated
from the type.

Squash;
 hiberman: Log in the event of a failed umount

 An ActiveMount when out of scope will attempt to umount
 the underlying volume. This could potentially fail however
 a Drop contract semantically does not encode runtime failure.

 Therefore, at least give us a log warn this happened and why
 until the control flow types have sufficient relationships
 to encode this runtime potentiality.

BUG=b:309654982
TEST=`emerge-nissa hiberman`.

Change-Id: I8bad835cf88902fec23316ff48df31adc962b0d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5119973
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>