multiple: Call correct FIPS vendor command
The current code calls FipsStatus, which only returns whether the FIPS
mode is active in GSC. It will return true even if the u2f state is
initialized with the legacy path.
A new subcommand is added to allow query for the U2F FIPS status
specifically. Switch to call that vendor command instead, and change
some namings to make the meanings clearer and more consistent.

In addition, add the commands to the fuzzed command list.

BUG=b:314250579
TEST=trunks/libhwsec unittests

Change-Id: I7a03e5f87f35eb8ac5b63c7b513fdba057b71678
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5118847
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>