cryptohome: Switch all USS access over to UssManager
This replaces all the copies of DecryptedUss in AuthSession with
accesses to the USS via UssManager instead. For the most part, it just
replaces most of the uses with a UssManager::GetDecrypted call, with
most of the more complex changes only being in the code which has to do
USS creation since that's a bit more involved.

This also replaces all the various bits of code that manually load the
EncryptedUss with lookups via UssManager as well.

The main purpose of this change is to ensure that all of the things
using the USS are in sync, using the same underlying data. But it should
also have the benefit of saving a bunch of redundant copies and extra
reloads of the USS from disk.

BUG=b:306423754
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ifbcf3c854f79eb98c996cb6832baa7c126ccffc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5124693
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>