clang-format: Use clang-format to sort includes
This allows clang-format to be used to sort #include lines, instead of
relying on cpplint which can tell you your ordering is wrong but not fix
it for you.

This is unfortunately not perfectly non-disruptive, as the precise way
in which cpplint liked to order things cannot be perfectly duplicated
using clang-format rules. The auto-grouping is also not prefect since it
has to rely on regexes for matching names, although in practice it
shouldn't be substantially worse than the pattern matching that cpplint
was doing.

After making this change and trying out cros format on a couple code
bases, there is a fair amount of noise. But a lot of that noise was from
code using incorrect ordering and/or grouping, so I'm not sure that's
a problem.

BUG=None
TEST=CQ

Change-Id: I431630cfb051e34c16fd2ba7f76cc7fe89919ee7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5124276
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>