missive: Resolve race in StorageQueue::Create.
The change makes it so that a queue instance is attempted to add
to the container before initializing, not after.
As a result:
- if the queue is new, it will start initialization as before.
- if the matching queue is already recorded, the process will drop the
new one and reuse the one that exists, and will schedule a callback once
the initialization has finished (successfully or not).

We will now never attempt to read the same directory more than once
in parallel and will never remove it upon failure to add to the container,
while another instance is trying to initialize the same directory.

Formatting change caused multiple trivial differences.
Added a test to validate race avoidance with 128 queue instance
being initialized in parallel.

BUG=b:315493850
TEST=cros_sdk cros_workon_make --board=${BOARD} missive --test
TEST=Manual VM run

Change-Id: I9bf703d1debb60d4fcb60467b1bccca22e7d9d16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5126119
Reviewed-by: Josh Hilke <jrhilke@google.com>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>