patchpanel: Disable bpf_loader temporarily
The recent change in kernel (https://crrev.com/c/5058379) changes the
output of mount options of bpffs, which in turns causes that the check
in security.Mtab test for the mount created by bpf_loader fails.

Since we're going to remove the mount() call in patchpanel_bpf_loader,
so instead of fixing it in the test now, let's disable the bpf_loader to
unblock the kernel patches to submit.

BUG=b:312117303
TEST=emerge, depoly patchpanel and reboot DUT, patchpanel is running
  after reboot.

Change-Id: Ieed22dc95ee8691d779938fcb2e18ff462238a87
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5124894
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>