shill: Fix key management during security upgrade
During security upgrade of an open network to transitional mode OWE and
pure OWE key management needs to change from "none" to "owe".  This
means that setting of key management can no longer be only set in
WiFiService constructor and needs to be updated during security
change/upgrade.

In addition to the above the handling of security update has been moved
to the UpdateSecurity().  This function is called on endpoints update
and this matters because:
- removal of endpoint can cause security change
- security property change of endpoint can also cause security update

BUG=b:315492221
TEST=FEATURES=test emerge-$BOARD shill
TEST=Either manually configure open network and then change it to pure
  OWE and verify that DUT connects or run tast test:
     tast run ... wifi.OweUpgrade
     from this CL: https://crrev.com/c/5106030

Change-Id: I248264a08e5ab8ebde5dbd6aff408f2b8dadefa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5126045
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>
Tested-by: Andrzej Ostruszka <andrzejo@google.com>
Auto-Submit: Andrzej Ostruszka <andrzejo@google.com>